% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsai_flathead.R
\name{bsai_flathead}
\alias{bsai_flathead}
\title{raw data query for BSAI flathead sole}
\usage{
bsai_flathead(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd, off_yr = NULL)
}
\arguments{
\item{year}{assessment year}

\item{akfin_user}{user name}

\item{akfin_pwd}{user password}

\item{afsc_user}{user name}

\item{afsc_pwd}{user password}

\item{off_yr}{if this is an off-year assessment change to TRUE}
}
\description{
raw data query for BSAI flathead sole
}
\examples{
\dontrun{
bsai_flathead(year = 2020, akfin_user, akfin_pwd, afsc_user, afsc_pwd)
}
}
