% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-ssid-errors.R
\name{correct_ssid_errors}
\alias{correct_ssid_errors}
\title{Custom fixes for problem surveys with shared trip ids resulting in assignment to wrong ssid}
\usage{
correct_ssid_errors(dat)
}
\arguments{
\item{dat}{df containing these columns: fishing_event_ids, survey_series_id, survey_id,
major_stat_area_code, minor_stat_area_code}
}
\description{
Custom fixes for problem surveys with shared trip ids resulting in assignment to wrong ssid
}
