% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-ssids.R
\name{correct_ssids}
\alias{correct_ssids}
\title{Custom fixes for problem surveys with shared trip ids resulting in assignment to wrong ssid}
\usage{
correct_ssids(dat, specimens = FALSE)
}
\arguments{
\item{dat}{df containing these columns: fishing_event_ids, survey_series_id, survey_id,
major_stat_area_code, minor_stat_area_code}

\item{specimens}{Defaults to FALSE where checks for duplication of fishing_event_ids}
}
\description{
Custom fixes for problem surveys with shared trip ids resulting in assignment to wrong ssid
}
