% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{design_indexes}
\alias{design_indexes}
\title{Design-based survey indexes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17400 rows and 8 columns.
}
\usage{
design_indexes
}
\description{
Design-based indexes as calculated from survey set data.
A bootstrap with 20,000 replicates is used for calculation
of standard errors and confidence intervals.
See file \code{data-raw/design-indexes.R} for calculation.
This is the output from running \code{\link[=get_design_index]{get_design_index()}}
on all the species in the groundfish data synopsis report.
}
\seealso{
\code{\link[=get_design_index]{get_design_index()}}
}
\keyword{datasets}
