% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent-level-counts.R
\name{get_parent_level_counts}
\alias{get_parent_level_counts}
\title{summarize sub/minor level counts from fishing event data at the parent event level
includes correction for a typo in dataframe
retrieves missing fishing_event_ids for sablefish surveys using major_level_ids}
\usage{
get_parent_level_counts(fe)
}
\arguments{
\item{fe}{df retrieved with get-event-data.sql}
}
\description{
summarize sub/minor level counts from fishing event data at the parent event level
includes correction for a typo in dataframe
retrieves missing fishing_event_ids for sablefish surveys using major_level_ids
}
