% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skate-level-counts.R
\name{get_skate_level_counts}
\alias{get_skate_level_counts}
\title{summarize fishing event data at the skate level
retrieves missing fishing_event_ids for sablefish surveys using major_level_ids}
\usage{
get_skate_level_counts(fe)
}
\arguments{
\item{fe}{df retrieved with get-event-data.sql}
}
\description{
summarize fishing event data at the skate level
retrieves missing fishing_event_ids for sablefish surveys using major_level_ids
}
