% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_spp_sample_length_type}
\alias{get_spp_sample_length_type}
\title{Length type}
\usage{
get_spp_sample_length_type(species)
}
\description{
Returns the most commonly recorded length measurement (of fork length, total
length, length to end of second dorsal fin, or standard length) for the
given species.
}
