% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-level-counts.R
\name{get_sub_level_counts}
\alias{get_sub_level_counts}
\title{extract sub/minor level counts from fishing event data
includes correction for a typo in dataframe
retrieves missing fishing_event_ids for sablefish surveys using major_level_ids}
\usage{
get_sub_level_counts(fe)
}
\arguments{
\item{fe}{df retrieved with get-event-data.sql}
}
\description{
extract sub/minor level counts from fishing event data
includes correction for a typo in dataframe
retrieves missing fishing_event_ids for sablefish surveys using major_level_ids
}
