% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfdata-package.R
\docType{package}
\name{gfdata-package}
\alias{gfdata}
\alias{gfdata-package}
\title{gfdata: Data queries for accessing the AKFIN and AFSC databases.}
\description{
This package provides a folder structure and SQL queries for accessing the AKFIN and AFSC databases in support of fishery stock assessments. Additionally, archived data are held/provided to maintain consistency across assessments/assessment authors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://benwilliams-noaa.github.io/gfdata}
}

}
\author{
\strong{Maintainer}: Benjamin Williams \email{ben.williams@noaa.gov}

}
\keyword{internal}
