% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_rebs.R
\name{goa_rebs}
\alias{goa_rebs}
\title{raw data query for GOA RE/BS rockfish}
\usage{
goa_rebs(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd, off_yr = NULL)
}
\arguments{
\item{year}{assessment year}

\item{akfin_user}{user name}

\item{akfin_pwd}{user password}

\item{afsc_user}{user name}

\item{afsc_pwd}{user password}

\item{off_yr}{if this is an off-year assessment change to TRUE}
}
\description{
raw data query for GOA RE/BS rockfish
}
