% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iphc_sets}
\alias{iphc_sets}
\alias{iphc_catch}
\title{IPHC data cleaned for use for non-halibut index standardization at PBS}
\format{
Once the 2 data frames are joined, a data frame with these columns:
\describe{
\item{year}{Year}
\item{station}{Station ID}
\item{station_key}{A unique station key from IPHC; occasionally there are multiple sets at the same \code{station} but this column is always unique. It's also unique across years.}
\item{longitude}{Longitude (mid point)}
\item{latitude}{Latitude (mid point)}
\item{species_science_name}{Scientific name}
\item{hooks_observed}{Number of hooks observed for non-halibut species}
\item{number_observed}{Number of hooks with the species of interest}
\item{pbs_standard_grid}{Logical: standard grid stations through time as defined in gfiphc}
\item{inside_wcvi}{Logical: inside Vancouver Island waters (2018 only) vs. anywhere else; you may want to exclude these from spatiotemporal modelling}
\item{sample_type}{Sample type (first 20 hooks vs. all hooks)}
\item{soak_time_min}{Soak time}
\item{temp_c}{Temperature in degrees C}
\item{depth_m}{Depth in m}
\item{species_common_name}{Species common name}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 238112 rows and 6 columns.
}
\usage{
iphc_sets

iphc_catch
}
\description{
These data are best used for spatiotemporal analyses. Also see the
\pkg{gfiphc} package which implements a design-based index standardization.
Data from 1998 onwards come from the IPHC website.
Data from 1996 and 1997 come from local spreadsheets as described in
\pkg{gfiphc}.
}
\details{
One likely wants to join the data frames. E.g.

\if{html}{\out{<div class="sourceCode">}}\preformatted{iphc <- dplyr::inner_join(iphc_catch, iphc_sets)
}\if{html}{\out{</div>}}
}
\keyword{datasets}
