% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lookup-tables.R
\name{lookup}
\alias{lookup}
\alias{get_ssids}
\alias{get_active_survey_blocks}
\alias{get_major_areas}
\alias{get_management_areas}
\alias{get_fishery_ids}
\alias{get_species_groups}
\alias{get_comm_gear_types}
\alias{get_survey_gear_types}
\alias{get_age_methods}
\alias{get_species}
\alias{get_strata_areas}
\alias{get_survey_ids}
\alias{get_sensor_attributes}
\alias{get_other_surveys}
\title{Lookup PBS metadata and descriptive data filters}
\usage{
get_ssids()

get_active_survey_blocks(ssid = NULL)

get_major_areas()

get_management_areas()

get_fishery_ids()

get_species_groups()

get_comm_gear_types()

get_survey_gear_types()

get_age_methods()

get_species()

get_strata_areas()

get_survey_ids(ssid)

get_sensor_attributes()

get_other_surveys()
}
\description{
Extracts metadata and descriptive details useful in filtering data (i.e.
for extracting data only from a specific survey). Codes are often used in
arguments or for filtering and can be looked up with these functions.
}
\details{
* `get_ssids()` produces a lookup table for survey series IDs and
   descriptions
* `get_age_methods()` produces a lookup table for ageing method codes
   and descriptions
* `get_sample_trips()`produces a lookup table for sample ID and
   fishing event ID
* `get_strata_areas()` produces a lookup table for surveyed area for
   each stratum within surveys
* `get_survey_ids()` produces lookup table for survey IDs for a given
   survey series ID
}
