% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_psc.R
\name{q_psc}
\alias{q_psc}
\title{query prohibited species catch estimate}
\usage{
q_psc(year, trip_target, area, akfin, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{trip_target}{'p' = pollock-mid, 'b' = pollock-bottom, 'x' = rex, 'h' = shallow flats, 'k' = rockfish, 'w' = arrowtooth, 'c' = pcod, 'i' = halibut}

\item{area}{goa or bsai}

\item{akfin}{the database to query}

\item{save}{save the file in designated folder}
}
\description{
prohibited species catch (PSC) estimates reported in tons for halibut and herring and counts for salmon and crabs and other fish
}
\examples{
\dontrun{

akfin = DBI::dbConnect(odbc::odbc(), "akfin", UID = akfin_user, PWD = akfin_pwd)
q_psc(year=2022, trip_target="k", area="goa", afkin, save=FALSE) \%>\%

}

}
