% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_sabl_length_comp.R
\name{q_sabl_length_comp}
\alias{q_sabl_length_comp}
\title{RPN weighted sablefish length comp data query}
\usage{
q_sabl_length_comp(year, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{save}{save the file in designated folder if false will load it to global env}
}
\value{

}
\description{
RPN weighted sablefish length comp data query
}
