% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_sabl_rpn.R
\name{q_sabl_rpn}
\alias{q_sabl_rpn}
\title{longline survey rpn/rpw data query}
\usage{
q_sabl_rpn(year, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{save}{save the file in designated folder if false will load it to global env}
}
\value{

}
\description{
longline survey rpn/rpw data query
}
