% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_shark_rpn.R
\name{q_shark_rpn}
\alias{q_shark_rpn}
\title{IPHC FISS RPN and CPUE from AKFIN}
\usage{
q_shark_rpn(year, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{save}{save the file in designated folder if false will load it to global env}
}
\description{
IPHC FISS RPN and CPUE from AKFIN
}
\examples{
\dontrun(q_sharl_rpn(year, save=FALSE))
}
