% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ts_app.R
\name{q_ts_app}
\alias{q_ts_app}
\title{trawl survey apportionment data query}
\usage{
q_ts_app(year, area = "goa", afsc_species, afsc, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{area}{default is "goa"}

\item{afsc_species}{afsc species code(s)}

\item{afsc}{the database to query}

\item{save}{save the file in designated folder}
}
\description{
trawl survey apportionment data query
}
