% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ts_biomass.R
\name{q_ts_biomass}
\alias{q_ts_biomass}
\title{trawl survey biomass data query}
\usage{
q_ts_biomass(
  year,
  area = "goa",
  afsc_species,
  shelf_slope = "slope",
  depth = TRUE,
  stratum = NULL,
  akfin,
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{area}{default is "goa"}

\item{afsc_species}{afsc species code(s)}

\item{shelf_slope}{if the area is the Bering Sea this default designated "shelf" alternate is "slope"}

\item{depth}{defaut is NULL, otherwise queries data by depth}

\item{stratum}{defaut is NULL, otherwise queries data by stratum}

\item{akfin}{the database to query}

\item{save}{save the file in designated folder, if FALSE outputs to global environment}
}
\value{

}
\description{
trawl survey biomass data query
}
