% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharks.R
\name{sharks}
\alias{sharks}
\title{query shark data}
\usage{
sharks(year, akfin_user, akfin_pwd, afsc_user, afsc_pwd, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{akfin_user}{user name}

\item{akfin_pwd}{user password}

\item{afsc_user}{user name}

\item{afsc_pwd}{user password}

\item{save}{save in standard location (default)}
}
\description{
query shark data
}
\examples{
\dontrun{
shakrs(year=2022, akfin_user, akfin_pwd, afsc_user, afsc_pwd, save = FALSE)
}
}
