% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_blocks}
\alias{survey_blocks}
\title{Active survey blocks}
\format{
Simple feature (\code{sf}) collection with 66744 features and 8 fields:
\describe{
\item{survey_abbrev}{Survey abbreviation.}
\item{survey_series_id}{Unique identifier for the survey series.}
\item{block_id}{Unique identifier for each grid cell (\code{BLOCK_DESIGNATION} in GFBioSQL).}
\item{grouping_code}{Strata grouping code used to join with strata data from
the GROUPING table in GFBioSQL.}
\item{depth_m}{Depth in metres.}
\item{active_block}{Is block actively fished as of date downloaded:
e.g., \code{attr(gfdata::survey_blocks, "date-downloaded")})}
\item{geometry}{Represents grid cell.}
\item{area}{Overwater area in km^2.}
}
}
\usage{
survey_blocks
}
\description{
Active survey blocks for DFO Pacific groundfish surveys.
Obtained via \code{gfdata::get_active_survey_blocks()} with some cleaning
as documented in \code{data-raw/survey_blocks.R}.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
requireNamespace("ggplot2", quietly = TRUE)
library(sf)
library(ggplot2)
gfdata::survey_blocks |>
  dplyr::filter(active_block) |>
  ggplot(aes(colour = survey_abbrev)) +
  geom_sf() +
  theme_minimal() +
  scale_colour_brewer(palette = "Dark2")
attr(gfdata::survey_blocks, "date-generated")
attr(gfdata::survey_blocks, "date-downloaded")
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
