% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dat.R, R/get-hake-dat.R, R/get-sablefish.R,
%   R/get-survey-samples.R, R/get-survey-sets.R
\name{get_data}
\alias{get_data}
\alias{get_survey_sets}
\alias{get_ll_hook_data}
\alias{get_survey_samples}
\alias{get_commercial_samples}
\alias{get_catch}
\alias{get_cpue_historical}
\alias{get_cpue_historical_hake}
\alias{get_cpue_historical_hl}
\alias{get_cpue_spatial}
\alias{get_catch_spatial}
\alias{get_cpue_spatial_ll}
\alias{get_cpue_index}
\alias{get_cpue_index_hl}
\alias{get_age_precision}
\alias{get_survey_index}
\alias{get_sable_landings}
\alias{get_survey_blocks}
\alias{get_eulachon_specimens}
\alias{get_gear_types}
\alias{get_management}
\alias{cache_pbs_data}
\alias{get_hake_survey_samples}
\alias{get_hake_catch}
\alias{get_sablefish_surveys}
\alias{get_survey_samples2}
\alias{get_survey_sets2}
\title{Get PBS data}
\usage{
get_survey_sets(
  species,
  ssid = c(1, 3, 4, 16, 2, 14, 22, 36, 39, 40),
  join_sample_ids = FALSE,
  verbose = FALSE,
  remove_false_zeros = FALSE,
  sleep = 0.05
)

get_ll_hook_data(species = NULL, ssid = NULL)

get_survey_samples(
  species,
  ssid = NULL,
  remove_bad_data = TRUE,
  unsorted_only = TRUE,
  usability = NULL,
  major = NULL
)

get_commercial_samples(
  species,
  unsorted_only = TRUE,
  return_all_lengths = FALSE,
  major = NULL,
  usability = NULL
)

get_catch(species, major = NULL)

get_cpue_historical(
  species = NULL,
  major = NULL,
  alt_year_start_date = "04-01",
  areas = c("3[CD]+", "5[AB]+", "5[CDE]+"),
  end_year = NULL
)

get_cpue_historical_hake(end_year = NULL)

get_cpue_historical_hl(
  species = NULL,
  major = NULL,
  alt_year_start_date = "04-01",
  areas = c("3[CD]+", "5[AB]+", "5[CDE]+"),
  end_year = NULL
)

get_cpue_spatial(species, major = NULL)

get_catch_spatial(species, major = NULL)

get_cpue_spatial_ll(species, major = NULL)

get_cpue_index(gear = "bottom trawl", min_cpue_year = 1996, major = NULL)

get_cpue_index_hl(min_cpue_year = 1980, major = NULL)

get_age_precision(species, major = NULL)

get_survey_index(species, ssid = NULL)

get_sable_landings(species, ssid = NULL)

get_survey_blocks(ssid = NULL)

get_eulachon_specimens()

get_gear_types()

get_management(
  species = NULL,
  species_group = NULL,
  fishery = NULL,
  area = NULL,
  start_year = NULL
)

cache_pbs_data(
  species,
  major = NULL,
  file_name = NULL,
  path = ".",
  compress = FALSE,
  unsorted_only = TRUE,
  historical_cpue = FALSE,
  survey_sets = FALSE,
  verbose = TRUE,
  return_all_lengths = FALSE
)

get_hake_survey_samples()

get_hake_catch(end_date = format(Sys.Date(), "\%d/\%m/\%Y"))

get_sablefish_surveys()

get_survey_samples2(
  species,
  ssid = NULL,
  remove_bad_data = TRUE,
  unsorted_only = TRUE,
  usability = NULL,
  major = NULL
)

get_commercial_samples(
  species,
  unsorted_only = TRUE,
  return_all_lengths = FALSE,
  major = NULL,
  usability = NULL
)

get_catch(species, major = NULL)

get_survey_sets2(
  species,
  ssid = c(1, 3, 4, 16, 2, 14, 22, 36, 39, 40),
  join_sample_ids = FALSE,
  verbose = FALSE,
  sleep = 0.05
)

get_ll_hook_data(species = NULL, ssid = NULL)
}
\arguments{
\item{species}{One or more species common names (e.g. \code{"pacific ocean perch"}) or one or more species codes (e.g. \code{396}). Species codes can be
specified as numeric vectors \verb{c(396, 442}) or characters \code{c("396", "442")}.
Numeric values shorter than 3 digits will be expanded to 3 digits and
converted to character objects (\code{1} turns into \code{"001"}). Species common
names and species codes should not be mixed. If any element is missing a
species code, then all elements will be assumed to be species common
names.}

\item{ssid}{A numeric vector of survey series IDs. Run \code{\link[=get_ssids]{get_ssids()}} for a
look-up table of available survey series IDs with surveys series
descriptions.}

\item{join_sample_ids}{If \code{TRUE} then the sample IDs will be joined in. This
may result in repeated rows of data if the same sample ID is part of
different survey stratifications.}

\item{verbose}{If \code{TRUE} then extra messages were reprinted during data
extraction. Useful to monitor progress.}

\item{remove_false_zeros}{If \code{TRUE} will make sure weights > 0 don't have
associated counts of 0 and vice versa. Only applies to trawl data where
counts are only taken for small catches.}

\item{sleep}{System sleep in seconds between each survey-year
to be kind to the server.}

\item{remove_bad_data}{Remove known bad data, such as unrealistic
length or weight values.}

\item{unsorted_only}{Remove sorted biological data ('keepers' and 'discards'
and unknown). Default = TRUE.}

\item{usability}{A vector of usability codes to include. Defaults to all.
IPHC codes may be different to other surveys.}

\item{major}{Character string (or vector, though doesn't work yet with
\code{cache_pbs_data}) of major stat area code to include (characters). Use
get_major_areas() to lookup area codes with descriptions.}

\item{return_all_lengths}{Include all length types, rather than just
with most common measurement. Default = FALSE.}

\item{alt_year_start_date}{Alternative year starting date specified as a
month-day combination. E.g. "03-01" for March 1st. Can be used to create
'fishing years'.}

\item{areas}{Area groupings as a vector of regular expressions.
See \code{\link[base:regex]{base::regex()}}.}

\item{end_year}{Specify the last calendar year to be extracted.}

\item{gear}{The gear type(s) to include for CPUE. Will be converted to
uppercase. Run \code{\link[=get_comm_gear_types]{get_comm_gear_types()}} for a look-up table of available
gear types to select from.}

\item{min_cpue_year}{Minimum year for the CPUE data.}

\item{species_group}{Species group code(s) to include (see lookup table
\code{\link[=get_species_groups]{get_species_groups()}}). Defaults to all.}

\item{fishery}{The fishery_id code(s) (see lookup table \code{\link[=get_fishery_ids]{get_fishery_ids()}})
for fisheries to include in data extraction. Defaults to all.}

\item{area}{The fishery area(s) (see lookup table \code{\link[=get_management_areas]{get_management_areas()}})
to include in data extraction (eg. '5A'; c('3C', '3D', '5A', '5B')).}

\item{start_year}{The minimum year to include management actions.
Defaults to all.}

\item{file_name}{Optional filename(s) for the cached file. Defaults to the
same as the \code{species} argument.}

\item{path}{The folder where the cached data will be saved.}

\item{compress}{Compress the \code{.rds} file? Defaults to \code{FALSE} for faster
reading and writing at the expense of disk space.}

\item{historical_cpue}{Logical for whether historical CPUE should be included.}

\item{survey_sets}{Logical for whether the survey set data should be
extracted. You might set this to \code{FALSE} if you don't need these data and
you want to substantially speed up data extraction.}

\item{end_date}{A string representing the date. Must be of the format dd/mm/yyyy}
}
\value{
The \verb{get_*} functions return a data frame. The \code{\link[=cache_pbs_data]{cache_pbs_data()}}
function writes an \code{.rds} file to \code{path} for each specified species. A data
object for a single species is a named list object with each element
containing a data frame from a \verb{get_*} function. The element name of the list
reflects the function name with the \code{get_} part removed. For example, the
output from \code{\link[=get_survey_samples]{get_survey_samples()}} is in a list element named
\code{survey_samples()}.
}
\description{
Automates fisheries and research survey data extraction from DFO Pacific
groundfish databases. The output datasets feed into other functions (\code{tidy_},
\code{plot_}, or \code{fit_} functions) for data visualization, which can be used as
products themselves or can be fed into automated DFO Pacific groundfish data
synopsis report production.
}
\details{
\itemize{
\item \code{get_survey_sets()} extracts survey catch data and spatial data for
plotting survey catches on a map of British Columbia
\item \code{get_survey_samples()} extracts all biological sample specimen records
from research surveys for given species and survey series IDs from GFBio
\item \code{get_hake_survey_samples()} extracts all biological sample specimen records
from hake joint acoustic surveys from GFBio
\item \code{get_commercial_samples()} extracts all biological sample specimen records
from commercial data for given species from GFBio
\item \code{get_catch()} extracts all landing and discard records for a given species
from GFFOS.GF_MERGED_CATCH
\item \code{get_hake_catch()} extracts all landing and discard records for Pacific Hake
with some extra data used in the Hake assessment
\item \code{get_cpue_spatial()} extracts catch, effort and spatial data from
GFFOS.GF_D_OFFICIAL_CATCH for the groundfish trawl fishery
\item \code{get_cpue_spatial_ll()} extracts catch, effort and spatial data from
GFFOS.GF_D_OFFICIAL_CATCH for the longline fishery
\item \code{get_cpue_index()} extracts catch and effort data from
GFFOS.GF_MERGED_CATCH for the groundfish trawl fishery since 1996
\item \code{get_cpue_historical()} extracts historical catch and effort data back into
the 1950s. It's help file is on a separate page; see the link
\item \code{get_age_precision()} extracts age readings from biological samples for a
given species where there is a second ('precision') age reading
\item \code{get_sara_dat()} scrubs Species At Risk website for up-to-date species
status and listings
\item \code{get_survey_index()} extracts survey catch data for given species
and survey series IDs
\item \code{get_management()} extracts management actions
\item \code{cache_pbs_data()} runs all 'get' functions in the gfdata package

(except those specific to IPHC data) and caches extracted data to a given
folder
}

This \code{\link[=cache_pbs_data]{cache_pbs_data()}} function caches data from
\itemize{
\item \code{\link[=get_survey_samples]{get_survey_samples()}}
\item \code{\link[=get_commercial_samples]{get_commercial_samples()}}
\item \code{\link[=get_catch]{get_catch()}}
\item \code{\link[=get_cpue_spatial]{get_cpue_spatial()}}
\item \code{\link[=get_cpue_spatial_ll]{get_cpue_spatial_ll()}}
\item \code{\link[=get_catch_spatial]{get_catch_spatial()}}
\item \code{\link[=get_survey_index]{get_survey_index()}}
\item \code{\link[=get_age_precision]{get_age_precision()}}
\item and optionally from \code{\link[=get_survey_sets]{get_survey_sets()}} and \code{\link[=get_cpue_historical]{get_cpue_historical()}}
}
}
\section{Authentication}{

\verb{get_*} functions only extract data when performed on a computer connected to
the Pacific Biological Station DFO network. By default, the functions assume
that you are on an authorized DFO Windows computer where authentication with
the databases happens automatically. If instead, you wish to connect by
username and password, see the details section in \code{\link[=run_sql]{run_sql()}}.
}

\examples{
\dontrun{
## Import survey catch density and location data by tow or set for plotting
## Specify single or multiple species by common name or species code and
## single or multiple survey series id(s).
## Notes:
## `area_km` is the stratum area used in design-based index calculation.
## `area_swept` is used to calculate density for trawl surveys and based on
## `area_swept1` (`doorspread_m` x `tow_length_m`) except when
## `tow_length_m` is missing, and then we use `area_swept2`
## (`doorspread` x `duration_min` x `speed_mpm`).
## `duration_min` is derived in the SQL procedure "proc_catmat_2011" and
## differs slightly from the difference between `time_deployed` and
## `time_retrieved`.

get_survey_sets(species = "lingcod", ssid = 1)

## Import survey or commercial biological data for various plots
## (e.g. length frequency, growth, age frequency, maturity, etc.)
get_survey_samples(species = 442, ssid = c(1, 3, 4, 16))

get_commercial_samples(c(442, 397))

## Import catch data by species for barcharts of landings by fishing area,
## geartype, and year.
get_catch("lingcod")

## Import spatial commercial catch per unit effort data for trawl or longline
## data by species for plotting along BC coast.
get_cpue_spatial("lingcod")
get_cpue_spatial_ll("yelloweye rockfish")

## Import catch and effort data by gear type for modelling commercial trawl
## cpue index.
get_cpue_index(gear = "bottom trawl", min_cpue_year = 2012)

## Import survey bootstrapped biomass estimates for plotting relative biomass
## indices by specified survey series.
get_survey_index("pacific cod", ssid = c(1, 3, 4, 16))
}

}
