% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-calc.R
\name{get_design_index}
\alias{get_design_index}
\title{Calculate a design index}
\usage{
get_design_index(species, ssid = NULL, reps = 1000)
}
\arguments{
\item{species}{Species common name or number}

\item{ssid}{Survey series ID}

\item{reps}{Number of bootstrap samples. Set to \code{0} to skip the bootstrapping
and only report design-based variance estimates.}
}
\description{
Calculate a design-based index with bootstrapped standard errors
}
