% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-envir-dat.R
\name{get_sensor_data_trawl}
\alias{get_sensor_data_trawl}
\alias{get_sensor_data_fe_trawl}
\alias{get_sensor_data_fe_ctd}
\alias{get_sensor_data_ll_td}
\title{Title}
\usage{
get_sensor_data_trawl(ssid = NULL, attribute = c("temperature",
  "depth", "dissolved oxygen", "salinity"), spread_attributes = FALSE)

get_sensor_data_fe_trawl(fishing_event_id = NULL,
  attribute = c("temperature", "depth", "dissolved oxygen", "salinity"),
  sensor_name = NULL)

get_sensor_data_fe_ctd(fishing_event_id = NULL,
  attribute = c("temperature", "depth", "dissolved oxygen", "salinity"),
  sensor_name = NULL)

get_sensor_data_ll_td(ssid = NULL, attribute = c("temperature",
  "depth"), sensor_min_max = FALSE)

get_sensor_data_ll_td(ssid = NULL, attribute = c("temperature",
  "depth"), sensor_min_max = FALSE)
}
\arguments{
\item{attribute}{A character vector of sensor attributes to filter for.
Run `get_sensor_attributes()` for a look-up table of available attributes.}

\item{spread_attributes}{Logical for whether the attributes should be
returned in a wider format.}

\item{fishing_event_id}{A vector of fishing events to filter for}

\item{sensor_name}{A character vector of sensor names to filter for.}

\item{sensor_min_max}{Allows for user to choose whether data are output in
wide format (= TRUE) with min and max values for each attribute for each
fishing event, or in long format (= FALSE) with only mean values for each
attribute and fishing event.}

\item{sensor_name}{TODO}
}
\description{
Title

Sensor data extraction by fishing event
}
\details{
To explore sensor data for an individual fishing event
}
