% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfdata-package.R
\docType{package}
\name{gfdata-package}
\alias{gfdata}
\alias{gfdata-package}
\title{gfdata: Data Extraction for DFO PBS Groundfish Stocks}
\description{
Facilitates groundfish data extraction at the Canadian Department of Fisheries and Oceans (DFO) Pacific Biological Station (PBS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pbs-assess/gfdata}
  \item Report bugs at \url{https://github.com/pbs-assess/gfdata/issues}
}

}
\author{
\strong{Maintainer}: Sean C. Anderson \email{sean.anderson@dfo-mpo.gc.ca}

Authors:
\itemize{
  \item Elise A. Keppel
  \item Andrew M. Edwards
  \item Chris Grandin
  \item Philina A. English
}

}
\keyword{internal}
