% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_dusk_catch_1977_1990}
\alias{goa_dusk_catch_1977_1990}
\title{GOA dusky rockfish historical catch data}
\format{
A data frame with 14 observations and 2 variables:
\describe{
\item{year}{year of catch}
\item{catch}{weight of catch in 1,000 t}
}
}
\usage{
goa_dusk_catch_1977_1990
}
\description{
A dataset containing trawl gear catch by year for 1977-1990
}
\keyword{datasets}
