% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mssm_grid}
\alias{mssm_grid}
\alias{mssm_grid_sf}
\title{Multispecies Small-Mesh Bottom Trawl (MSSM) Survey Grid}
\format{
\subsection{\code{mssm_grid}}{

A data frame with 3,735 rows and 5 columns:
\describe{
\item{survey_abbrev}{Survey abbreviation}
\item{longitude, latitude}{Longitude and latitude of the centroid of 3x3 km
grid cells}
\item{area}{Area of the grid cells, in km^2}
\item{year}{The year a grid cell was sampled. A grid cell can have multiple year values}
}
}

\subsection{\code{mssm_grid_sf}}{

A simple features (\code{sf} object) version of \code{mssm_grid}
}

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3735 rows and 3 columns.
}
\usage{
mssm_grid

mssm_grid_sf
}
\description{
A 3x3 km grid for the Multispecies Small-Mesh Bottom Trawl Survey (MSSM; formerly
known as 'shrimp survey'). This grid covers WCVI Shrimp Survey Areas 124 and 125.
The \code{year} is the last year a grid cell was sampled as of 2023, with
the most consistent resampling (spatially) occurring in grid cells last sampled
between 2009 and 2021.
}
\seealso{
\code{\link{load_survey_blocks}} for loading this dataset with additional options
(polygon, centroid, or coordinate formats).
}
\keyword{datasets}
