% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_blocks}
\alias{survey_blocks}
\title{Active survey blocks}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 22026 rows and 7 columns.
}
\usage{
survey_blocks
}
\description{
Active survey blocks for DFO Pacific groundfish surveys.
Obtained via \code{gfdata::get_active_survey_blocks()} with some cleaning
as documented in \code{data-raw/survey_blocks.R}.
}
\keyword{datasets}
