% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_blocks}
\alias{survey_blocks}
\title{Synoptic and HBLL survey blocks}
\format{
Simple feature (\code{sf}) collection with 66744 features and 7 fields:
\describe{
\item{survey_abbrev}{Survey abbreviation. Includes:
\itemize{
\item \code{SYN QCS}: Synoptic Queen Charlotte Sound
\item \code{SYN HS}: Synoptic Hecate Strait
\item \code{SYN WCVI}: Synoptic West Coast Vancouver Island
\item \code{SYN WCHG}: Synoptic West Coast Haida Gwaii
\item \code{HBLL OUT N}: Hook and Line Outside North
\item \code{HBLL OUT S}: Hook and Line Outside South
\item \code{HBLL INS N}: Hook and Line Inside North
\item \code{HBLL INS S}: Hook and Line Inside South
}
}
\item{survey_series_id}{Unique identifier for the survey series.}
\item{block_id}{Unique identifier for each grid cell (\code{BLOCK_DESIGNATION} in GFBioSQL).}
\item{grouping_code}{Strata grouping code used to join with strata data from
the GROUPING table in GFBioSQL.}
\item{depth_m}{Depth in metres.}
\item{active_block}{Is block actively fished as of date downloaded:
e.g., \code{attr(gfdata::survey_blocks, "date-downloaded")})}
\item{geometry}{Represents grid cell.}
\item{area}{Overwater area in km^2.}
}
}
\usage{
survey_blocks
}
\description{
Survey blocks (grids) for DFO Pacific groundfish surveys including Synoptic (SYN) and
Hook and Line (HBLL) surveys. Obtained via \code{gfdata::get_active_survey_blocks()}
with some cleaning as documented in \code{data-raw/survey_blocks.R}.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
requireNamespace("ggplot2", quietly = TRUE)
library(sf)
library(ggplot2)
gfdata::survey_blocks |>
  dplyr::filter(active_block) |>
  ggplot(aes(colour = survey_abbrev)) +
  geom_sf() +
  theme_minimal() +
  scale_colour_brewer(palette = "Dark2")
attr(gfdata::survey_blocks, "date-generated")
attr(gfdata::survey_blocks, "date-downloaded")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link{load_survey_blocks}} for loading this dataset with additional options
(polygon, centroid, or coordinate formats).
}
\keyword{datasets}
